Atualização 0.2.1
Redefinindo o basico, gerando arvores de forma aleatória, definindo as classes num arquivo separado para facilitar o uso, criando varios mapas (todas as definições e o esqueleto estão aqui, faltam só desenhar os mapas e inserir no jogo), melhorando várias definições problemáticas anteriormente.


Bug persistente - Ao encostar em certos sprites, o personagem e as aranhas podem teleportar. Precisarei refazer a colisão para corrigir, provavelmente.
