+++ Journey to fairy city/var.py
@@ -1,4 +1,19 @@
-import random
+import pygame, random, sys
+
+pygame.init()
+pygame.font.init()
+
+screen_width = 800 #largura
+screen_height = 600 #altura
+tela = pygame.display.set_mode((screen_width,screen_height))
+
+# Define as cores utilizadas no jogo
+white = (255, 255, 255)
+black = (0, 0, 0)
+green = (0, 255, 0)
+red = (255, 0, 0)
+
+#variaveis do jogo
 
 top = False
 bottom = False
@@ -16,7 +31,28 @@ machadinho = True
 menu = True
 game = False
 run_game = False
+posinitix = screen_width / 2
+posinitiy = screen_height / 2
+score = 0
+money = 0
 
+#variaveis do mapa
+Norte = 0
+Sul = 0
+Leste = 0
+Oeste = 0
+
+fase_atual = 'FN'
+
+#grupos de sprites
+attack_grupo = pygame.sprite.Group()
+arvore_grupo = pygame.sprite.Group()
+monstro_grupo = pygame.sprite.Group()
+char_grupo = pygame.sprite.Group()
+borda_grupo = pygame.sprite.Group()
+
+#fontes pro HUD
+font = pygame.font.SysFont('Arial', 30)
 
 def randomgen():
     randomizando = random.randrange(0,10)
